refactor: README.md 수정 등
 - Job contents :
  -. README.md 수정
  -. Commit Template 추가
  -. 몇몇 파일 이름 변경
  -. 테스트 코드 수정

 - Why did you do it? :
  -. 지속적인 리팩토링의 일환
