+++ .github/ISSUE_TEMPLATE/issue.md
@@ -1,0 +1,16 @@
+---
+
+|about|title|labels|assignees|
+|---|---|---|---|
+|이슈 템플릿의 목적|제목|라벨|담당자|
+
+---
+
+## 주요 문제점
+
+---
+
+## 의견
+
+---
+
