+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,1 @@
+미정
