+++ src/main/java/toy/subscribe/mvc/repository/FeedBoardCustomRepository.java
@@ -2,10 +2,10 @@ package toy.subscribe.mvc.repository;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import toy.subscribe.domain.dto.FeedBoardDto;
+import toy.subscribe.domain.response.FeedBoardResponseDto;
 
 public interface FeedBoardCustomRepository {
     
-    Page<FeedBoardDto> findPageByFeedBoard(Pageable pageable, String company, String title);
+    Page<FeedBoardResponseDto> findPageByFeedBoard(Pageable pageable, String company, String title);
     
 }
