+++ src/main/java/toy/subscribe/mvc/repository/FeedBoardCustomRepositoryImpl.java
@@ -1,18 +1,19 @@
 package toy.subscribe.mvc.repository;
 
-import static toy.subscribe.domain.entity.QFeedBoard.feedBoard;
-
 import com.querydsl.core.QueryResults;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import java.util.List;
-import java.util.stream.Collectors;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.support.QuerydslRepositorySupport;
 import org.springframework.transaction.annotation.Transactional;
-import toy.subscribe.domain.dto.FeedBoardDto;
 import toy.subscribe.domain.entity.FeedBoard;
+import toy.subscribe.domain.response.FeedBoardResponseDto;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+import static toy.subscribe.domain.entity.QFeedBoard.feedBoard;
 
 public class FeedBoardCustomRepositoryImpl extends QuerydslRepositorySupport implements FeedBoardCustomRepository {
     
@@ -22,7 +23,7 @@ public class FeedBoardCustomRepositoryImpl extends QuerydslRepositorySupport imp
     
     @Override
     @Transactional(readOnly = true)
-    public Page<FeedBoardDto> findPageByFeedBoard(Pageable pageable, String company, String title) {
+    public Page<FeedBoardResponseDto> findPageByFeedBoard(Pageable pageable, String company, String title) {
         
         JPAQueryFactory queryFactory = new JPAQueryFactory(getEntityManager());
         QueryResults<FeedBoard> results = queryFactory
@@ -34,12 +35,12 @@ public class FeedBoardCustomRepositoryImpl extends QuerydslRepositorySupport imp
                 .limit(pageable.getPageSize())
                 .orderBy(feedBoard.id.desc())
                 .fetchResults();
-
-        List<FeedBoardDto> content = results.getResults()
-                .stream()
-                .map(FeedBoardDto::convertToFeedBoardDtoFrom)
-                .collect(Collectors.toList());
-
+        
+        List<FeedBoardResponseDto> content = results.getResults()
+                                                    .stream()
+                                                    .map(FeedBoardResponseDto::convertToFeedBoardDtoFrom)
+                                                    .collect(Collectors.toList());
+        
         return new PageImpl<>(content, pageable, results.getTotal());
     }
